Refactoring.
Keep external access (database and web calls) siloed in minimal functions. Make the rest of the code consume pre-processed data. This makes the tests easier to write (fewer functions that need mocking) and it also gives a speed benefit in that we only look up a given item one time instead of once for every instance of that item in the auction house data.