Get WoW running again
There is an item #2460 for which Blizzard returns an error when you look it up. This exposed some holes in the error handling. Update the handling to be more resilient.

The -clientID and -clientSecret flags are required. Check for them and issue a usage message if they are not provided.