Better item handling
Download and store the entire item descriptor. Use accessor functions to
read from the descriptor. This is far less fragile than the previous
approach, which was to compute the values and throw away the descriptor.
This meant if there was an error in the computed value we had to
invalidate the cache and start over.